refactor: 확장함수 이동
